More JEI related stuff
this should have all except for the nasa workbench mostly functional.

The textures still need updating to different ones if needed, for now they are just using the single texture.

This supports the fluid tanks draining/filling, the energy tanks draining/filling, using the same size as the standards in the inventory screens