Re-work of planet sky rendering
This does the following:

Renders planets and moons in the sky.
If it is rendering a moon, and is set as tidally locked, it renders the parent body in a fixed location based on the x-coordinate.
Moons of the current planet will render, and if they have a phase texture, it will use that, with a phase based on relative angle of the sun and the moon.
Planets will move throughout the sky depending on their location in orbit around the star.
Other moons will also appear in the sky if you are on a Moon

Configurable planet file can be used to adjust things like:

Whether clouds render on the planet.
Whether rain renders on the planet.
Offsets for sunrise/sunset colours